efinet: efinet_dev_init
../efinet.c: In function 'efinet_dev_init':
../efinet.c:315:46: error: format '%d' expects argument of type 'int', but argument 2 has type 'UINTN {aka long unsigned int}' [-Werror=format=]
    printf("Unable to open network interface %d for "
                                             ~^
                                             %ld
cc1: all warnings being treated as errors

+++ usr/src/boot/sys/boot/efi/libefi/efinet.c
@@ -313,7 +313,7 @@ efinet_dev_init()
 		    IH, NULL, EFI_OPEN_PROTOCOL_EXCLUSIVE);
 		if (status != EFI_SUCCESS) {
 			printf("Unable to open network interface %d for "
-			    "exclusive access: %lu\n", i,
+			    "exclusive access: %lu\n", (int)i,
 			    EFI_ERROR_CODE(status));
 		}
